feat: reorganize project structure
