+++ projects/02_reading_list/internal/api/data/books-model.go
